+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -762,6 +762,7 @@ class ConfigurationTest extends TestCase
                     'localizable_html_attributes' => [],
                 ],
                 'providers' => [],
+                'globals' => [],
             ],
             'validation' => [
                 'enabled' => !class_exists(FullStack::class),